add another exported share component
