+++ src/containers/products/CareProductPage.js
@@ -19,18 +19,19 @@ import { Content, PageDescriptionWrapper } from '../../components/layout/IntroCo
 import {
   ExtendedFeatureDescription,
   ExtendedFeatureHeader,
+  ExtendedFeaturesContent,
   FeatureOverviewDescription,
   FeatureOverviewTitle,
   FeatureOverviewWrapper,
   FeatureShot,
   Icon,
+  ProductFeaturesContent,
   ProductOverviewDescriptionWrapper,
   ProductOverviewWrapper,
   ProductShot,
   ProductStamp,
 } from './styled/StyledProductComponents';
 import {
-  CONTENT_WIDTH,
   MEDIA_QUERY_LG,
   MEDIA_QUERY_MD,
   MEDIA_QUERY_TECH_SM,
@@ -73,46 +74,9 @@ const CareProductShot = styled(ProductShot)`
   }
 `;
 
-const ProductFeaturesContent = styled.div`
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-
-  @media only screen and (max-width: ${MEDIA_QUERY_MD}px) {
-    margin-top: 300px;
-  }
-
+const CAREFeaturesContent = styled(ProductFeaturesContent)`
   @media only screen and (min-width: ${MEDIA_QUERY_LG}px) {
-    align-items: flex-start;
-    flex-direction: row;
     margin: 391px auto 96px auto;
-    width: ${CONTENT_WIDTH}px;
-  }
-
-  @media only screen and (max-width: ${MEDIA_QUERY_TECH_SM}px) {
-    margin-top: 150px;
-  }
-`;
-
-const ExtendedFeaturesContent = styled(ProductFeaturesContent)`
-  align-self: center;
-  margin-top: 0;
-
-  @media only screen and (max-width: ${MEDIA_QUERY_MD}px) {
-    margin-top: 32px;
-  }
-
-  @media only screen and (min-width: ${MEDIA_QUERY_LG}px) {
-    align-items: flex-start;
-    flex-direction: column;
-    margin: 0 0 200px 0;
-    width: ${CONTENT_WIDTH}px;
-  }
-
-  @media only screen and (max-width: ${MEDIA_QUERY_TECH_SM}px) {
-    margin: 40px 0;
   }
 `;
 
@@ -134,7 +98,7 @@ const CareProductPage = () => (
       <CareProductShot bgImage={ProductImg} />
     </PageSection>
     <PageSection>
-      <ProductFeaturesContent>
+      <CAREFeaturesContent>
         <FeatureOverviewWrapper>
           <img src={Feature1} alt="" />
           <FeatureOverviewTitle>Dispatch integration</FeatureOverviewTitle>
@@ -159,7 +123,7 @@ const CareProductPage = () => (
               response plans for frequently-encountered individuals.`}
           </FeatureOverviewDescription>
         </FeatureOverviewWrapper>
-      </ProductFeaturesContent>
+      </CAREFeaturesContent>
     </PageSection>
     <PageSection>
       <ExtendedFeaturesContent>
